Move tasks into files
