+++ src/identity_committer/tasks/mod.rs
@@ -1,0 +1,3 @@
+pub mod fetch_and_enqueue_pending_identities;
+pub mod mine_identities;
+pub mod process_identities;
